Headless mode and refactor
